[chore] Feed 데이터 클래스 재활용을 위한 네이밍 수정 (#91)
