+++ domain/src/main/java/com/readme/android/domain/entity/response/DomainHomeFeedResponse.kt
@@ -1,8 +1,8 @@
 package com.readme.android.domain.entity.response
 
-import com.readme.android.domain.entity.Feed
+import com.readme.android.domain.entity.FeedInfo
 
 data class DomainHomeFeedResponse(
     val filters: List<String>,
-    val feeds: List<Feed>
+    val feedInfos: List<FeedInfo>
 )
