+++ domain/src/main/java/com/readme/android/domain/repository/FeedRepository.kt
@@ -5,4 +5,6 @@ import com.readme.android.domain.entity.response.DomainHomeFeedResponse
 interface FeedRepository {
 
     suspend fun getHomeFeed(filters: String): Result<DomainHomeFeedResponse>
+
+    // suspend fun getDetailFeed(int: Int): Result<>
 }
