+++ feature/book-search/src/main/java/com/readme/android/book_search/BookListRecyclerViewAdapter.kt
@@ -1,17 +1,14 @@
 package com.readme.android.book_search
 
 import android.os.Build
-import android.text.Html
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.annotation.RequiresApi
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.readme.android.book_search.databinding.ItemBookSearchBinding
 import com.readme.android.core_ui.util.ItemDiffCallback
 import com.readme.android.domain.entity.BookInfo
-import com.readme.android.domain.entity.Feed
 import com.readme.android.navigator.MainNavigator
