+++ feature/main/src/main/res/layout/activity_feed_detail.xml
@@ -9,6 +9,7 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
+        android:background="@color/gray_0"
         android:layout_height="match_parent"
         tools:context=".ui.feed.FeedDetailActivity">
 
@@ -48,6 +49,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="32dp"
+                    app:layout_constraintBottom_toBottomOf="parent"
                     android:background="@drawable/shape_rect_white_fill_top_right_30"
                     app:layout_constraintTop_toBottomOf="@id/layout_book_information">
