+++ feature/main/src/main/res/layout/item_feed.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="feedData"
-            type="com.readme.android.domain.entity.Feed" />
+            type="com.readme.android.domain.entity.FeedInfo" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
